Rearrange implementations
