+++ backend/mirai-console/src/main/kotlin/net/mamoe/mirai/console/util/Annotations.kt
